requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
python-dotenv>=0.19.0

# NLP and sentiment analysis
nltk>=3.6.5
transformers>=4.11.0
torch>=1.9.0
numpy>=1.21.0
vaderSentiment>=3.3.2

# Google Gemini integration
google-generativeai>=0.3.0

# Utilities
python-multipart>=0.0.5
email-validator>=1.1.3
aiofiles>=0.7.0
python-jose[cryptography]>=3.3.0
httpx>=0.23.0

# Testing
pytest>=6.2.5
pytest-asyncio>=0.16.0 